Create PlayerTracker
The way this works is it simply tracks the x-coordinates of the object it is attached to and compares them to a pre-defined set of x-coordinates, in this case the "finish line" and the "checkpoint". The coordinates I used for finish1, finish2, checkPoint1, and checkPoint2 were just coordinates that worked for the little arena that I was testing this in, so once we have an actual track they can obviously be changed to visually match where the finish line is at along with any checkpoints we have. In this case the checkpoint wasn't even made known to the user, but rather as a way to ensure that the play actually went around the track and didn't cheat the system.